Refactor #321 - cleanup store impl files
