--- statekit/core/src/main/kotlin/com/nlab/statekit/store/DefaultStoreFactory.kt
@@ -1,48 +1,0 @@
-/*
- * Copyright (C) 2024 The N's lab Open Source Project
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *        http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.nlab.statekit.store
-
-import com.nlab.statekit.bootstrap.Bootstrap
-import com.nlab.statekit.reduce.AccumulatorPool
-import com.nlab.statekit.reduce.DefaultActionDispatcher
-import com.nlab.statekit.reduce.Reduce
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.asStateFlow
-
-/**
- * @author Doohyun
- */
-internal class DefaultStoreFactory {
-    private val accPool = AccumulatorPool()
-
-    fun <A : Any, S : Any> createStore(
-        coroutineScope: CoroutineScope,
-        initState: S,
-        reduce: Reduce<A, S>,
-        bootstrap: Bootstrap<A>
-    ): Store<A, S> {
-        val baseState = MutableStateFlow(initState)
-        val actionDispatcher = DefaultActionDispatcher(reduce, baseState, accPool)
-        return DefaultStore(
-            baseState.asStateFlow(),
-            coroutineScope,
-            actionDispatcher,
-            initJobs = bootstrap.fetch(coroutineScope, actionDispatcher, baseState.subscriptionCount)
-        )
-    }
-}
