+++ statekit/core/src/test/kotlin/com/nlab/statekit/store/StoreFactoryTest.kt
@@ -19,11 +19,11 @@ import org.mockito.kotlin.verify
 /**
  * @author Doohyun
  */
-class DefaultStoreFactoryTest {
+class StoreFactoryTest {
     @Test
     fun `Given initState, When create store, Then store has initState`() = runTest {
         val initState = TestState.genState()
-        val store = createStoreFromDefaultStoreFactory(initState = initState)
+        val store = createStoreFromStoreFactory(initState = initState)
 
         assertThat(store.state.value, equalTo(initState))
     }
@@ -32,18 +32,18 @@ class DefaultStoreFactoryTest {
     fun `When create store, Then bootstrap fetched`() = runTest {
         val initState = TestState.genState()
         val bootstrap: Bootstrap<TestAction> = mock()
-        createStoreFromDefaultStoreFactory(initState = initState, bootstrap = bootstrap)
+        createStoreFromStoreFactory(initState = initState, bootstrap = bootstrap)
 
         verify(bootstrap, once()).fetch(any(), any(), any())
     }
 }
 
-private fun TestScope.createStoreFromDefaultStoreFactory(
+private fun TestScope.createStoreFromStoreFactory(
     coroutineScope: CoroutineScope = this,
     initState: TestState = TestState.genState(),
     reduce: Reduce<TestAction, TestState> = Reduce(),
     bootstrap: Bootstrap<TestAction> = EmptyBootstrap()
-): Store<TestAction, TestState> = DefaultStoreFactory().createStore(
+): Store<TestAction, TestState> = StoreFactory().createStore(
     coroutineScope,
     initState,
     reduce,
